pyproject.toml
[project]
name = "guillemot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "logfire>=4.6.0",
    "pydantic-ai>=1.0.3",
    "python-dotenv>=1.1.1",
]
